pyproject.toml
[tool.poetry]
name = "otuserver_low_level"
version = "0.1.0"
description = ""
authors = ["ibukin <ibukin@voice-link.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dateutil = "^2.8.2"
pytest = "^7.1.3"
redis = "^4.3.4"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
